meta-ads-bot/ghost-mouse/requirements.txt
# ══════════════════════════════════════════════
# GHOST MOUSE — Dependencies
# ══════════════════════════════════════════════
playwright==1.52.0
playwright-stealth==1.0.6
websockets==15.0.1
aiohttp==3.11.18
anthropic==0.52.0
python-dotenv==1.1.0
bezier==2024.6.30
Pillow==11.2.1
